Run integrity tests from inside the container (#27)
* Run integrity tests from inside the container

Current flow is to run testinfra from host machine. The issue
with this approach is that overhead on communicating with the
test container is too high: basically every command in the test
needs to go through ansible and this cripples tests performance
on big packages. New approach is to mount tests into test container
and run testinfra from inside it.

* Fixed tests to go over all files in a package

* Fixed comments from review

* Fixed test if package was installed

Co-authored-by: Vasily Kleschov <vkleschov@cloudlinux.com>